Move GUI parts into `main/gui/`
